Move nodes getting to section component
